requirements.txt
flask==3.0.2
firebase-admin==6.4.0
python-dotenv==1.0.1
requests>=2.32.3
deepseek-ai==0.0.1
gunicorn==21.2.0
flask-cors==4.0.0
python-dateutil==2.8.2
schedule==1.2.1 